feat: move version constant to a new version.rb file and update gemspec references
