+++ lib/deadfinder.rb
@@ -8,7 +8,7 @@ require 'deadfinder/utils'
 require 'deadfinder/logger'
 require 'deadfinder/runner'
 require 'deadfinder/cli'
-require 'deadfinder/const'
+require 'deadfinder/version'
 require 'concurrent-edge'
 require 'sitemap-parser'
 require 'json'
