+++ lib/deadfinder/version.rb
